Include d3dx9 tools in module
