+++ Common/LoadModules.h
@@ -2,4 +2,5 @@
 
 void LoadASIPlugins(bool LoadFromScriptsOnly);
 void InitializeASI(HMODULE hModule);
+void ExtractD3DX9Tools();
 HRESULT DeleteAllfiles(LPCWSTR lpFolder);
