+++ Resources/BuildNo.rc
@@ -1,1 +1,1 @@
-#define BUILD_NUMBER 1859 
+#define BUILD_NUMBER 1860
